# MINOR: Update to Gradle 6.5 and tweak build jvm config