--- a/build.gradle
+++ b/build.gradle
@@ -506,8 +506,8 @@ subprojects {
       scalaCompileOptions.additionalParameters += ["-release", minJavaVersion]
 
     configure(scalaCompileOptions.forkOptions) {
-      memoryMaximumSize = '1g'
       jvmArgs = ['-Xss4m']
+      memoryMaximumSize = '2g'
     }
   }