@ijuma 2020-06-03
JDK 8/Scala 2.12 had an out of memory issues.