@ijuma 2020-06-03
Cherry picked this to 2.6 as well since it's low risk and fixes a compilation speed regression.